+++ firmware/main/uni_bluetooth.c
@@ -718,9 +718,6 @@ static void on_l2cap_data_packet(uint16_t channel, uint8_t* packet,
     return;
   }
 
-  printf("on_l2cap_data_packet\n");
-  printf_hexdump(packet, size);
-
   if (channel != device->hid_interrupt_cid) return;
 
   if (!uni_hid_device_has_hid_descriptor(device)) {