+++ firmware/main/uni_hid_parser_wiiupro.h
@@ -16,27 +16,19 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ****************************************************************************/
 
-#include "uni_bluetooth.h"
-#include "uni_hid_device.h"
-#include "uni_platform.h"
+#ifndef UNI_HID_PARSER_WIIUPRO_H
+#define UNI_HID_PARSER_WIIUPRO_H
 
-int btstack_main(int argc, const char** argv);
+#include <stdint.h>
 
-// Main. Called by BlueKitchen bluetooth stack
-int btstack_main(int argc, const char** argv) {
-  UNUSED(argc);
-  UNUSED(argv);
+#include "btstack.h"
 
-  // Honoring with BT copyright + adding own message to avoid confusion
-  printf("Unijoysticle 2 (C) 2016-2019 Ricardo Quesada and contributors.\n");
-  printf("Bluetooth stack: Copyright (C) 2017 BlueKitchen GmbH.\n");
-  printf("Firmware version: v0.2.1\n");
+#include "uni_gamepad.h"
+#include "uni_hid_parser.h"
 
-  uni_platform_init();
-  uni_hid_device_init();
+// For the Nimbus gamepad.
+void uni_hid_parser_wiiupro_init(uni_gamepad_t* gp);
+void uni_hid_parser_wiiupro_parse_raw(uni_gamepad_t* gp, const uint8_t* report,
+                                      uint16_t len);
 
-  // Continue with bluetooth setup.
-  uni_bluetooth_init();
-
-  return 0;
-}
+#endif  // UNI_HID_PARSER_WIIUPRO_H