+++ CHANGELOG.md
@@ -4,14 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.1] - 2019-06-29
 ### Added
 - Board: moved SparkFun.lib inside unijoysticle project
 - Firmware: more verbose logs when detecting the type of device
+- Firmware: Started Wii U Pro controller support. Not working yet.
 
 ### Changed
 - Firmware: Gamepad names are fetched correctly.
 - Firmware: Using btstack develop-branch. Commit: 32b46fec1df77000b2e383d209074f4c2866ebdf
+- Firmware: "apple" parser renamed to "nimbus" parser.
 
 ## [0.2.0] - 2019-05-22
 ### Added